requirements-ocr.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
rapidocr-onnxruntime==1.2.3
Pillow==11.3.0
numpy==2.2.6
mlx-lm==0.31.2
mlx-whisper==0.4.3

# Optional: if runtime supports Paddle, service will auto-switch to paddle backend.
# paddleocr==2.10.0
# paddlepaddle==3.1.1
